requirements.txt
streamlit
speechrecognition
gtts
langdetect
deep-translator
pillow
numpy
pandas
scikit-learn
joblib
requests
tk